+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.90.18
+- Added service type gRPC, you can now check on gRPC services. (limited)
+-
+
 # 0.90.17
 - Fixed notification fields for frontend
 - Fixed notification JSON form to send integer if value is an integer.
@@ -23,7 +27,7 @@
 # 0.90.13
 - Added new function `utils.RenameDirectory` to rename directory
 - Added new function `(*DbConfig) BackupAssets` to backup a customized theme and place into a directory named `assets_backup`. Only for migration 0.80 to 0.90+, entirely new frontend.
-- Updated JS function `convertToChartData` to return an empty chart data if API response was empty. 
+- Updated JS function `convertToChartData` to return an empty chart data if API response was empty.
 - Updated `banner.png` to make a bit smaller, (680px)
 - Fixed method that returns `no such table: services` on startup, check table first.
 - Fixed version from not being added into Core table. [Issue #436](https://github.com/statping/statping/issues/436)